+++ x/perp/v2/module/genesis.go
@@ -74,7 +74,7 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	}
 
 	if genState.CollateralDenom != "" {
-		err := k.Admin.UnsafeChangeCollateralDenom(ctx, genState.CollateralDenom)
+		err := k.Sudo().UnsafeChangeCollateralDenom(ctx, genState.CollateralDenom)
 		if err != nil {
 			panic(err)
 		}